+++ joplin/base/models/__init__.py
@@ -25,6 +25,7 @@ from .janis_page import JanisPage, JanisBasePage
 from .home_page import HomePage
 from .service_page import ServicePage
 from .translated_image import TranslatedImage
+from .information_page import InformationPage
 
 WYSIWYG_GENERAL = ['h1', 'h2', 'h3', 'h4', 'bold', 'link', 'ul', 'ol', 'code']
 DEFAULT_MAX_LENGTH = 255
@@ -63,45 +64,6 @@ class ProcessPage(JanisPage):
         ImageChooserPanel('image'),
         InlinePanel('contacts', label='Contacts'),
         InlinePanel('process_steps', label="Process steps"),
-    ]
-
-class InformationPage(JanisPage):
-    janis_url_page_type = "information"
-
-    description = models.TextField(blank=True, verbose_name='Write a description of this page')
-    options = StreamField(
-        [
-            ('option', RichTextBlock(
-                features=WYSIWYG_GENERAL,
-                label='Option'
-            ))
-        ],
-        verbose_name='Add option sections as needed.',
-        help_text='Options are needed when the reader needs to make a choice between a few options, such as ways to fill out a form (online, by phone, in person, etc.).',
-        blank=True
-    )
-
-    additional_content = RichTextField(
-        features=WYSIWYG_GENERAL,
-        verbose_name='Write any additional content describing the service',
-        blank=True
-    )
-
-    # TODO: Add images array field
-
-    base_form_class = custom_forms.InformationPageForm
-
-    content_panels = [
-        FieldPanel('title_en'),
-        FieldPanel('title_es'),
-        FieldPanel('title_ar'),
-        FieldPanel('title_vi'),
-        InlinePanel('topics', label='Topics'),
-        InlinePanel('related_departments', label='Related Departments'),
-        FieldPanel('description'),
-        StreamFieldPanel('options'),
-        FieldPanel('additional_content'),
-        InlinePanel('contacts', label='Contacts'),
     ]
 
 class TopicCollectionPage(JanisPage):
